+++ src/com/mcmorrow/bank/app/Savings.java
@@ -5,7 +5,7 @@ public class Savings extends Account {
     private int safetyDepositBoxKey;
 
 
-    public Savings(String name, String sSN, int initDeposit) {
+    public Savings(String name, String sSN, double initDeposit) {
         super(name, sSN, initDeposit);
         accountNumber = "1"+accountNumber;
         setSafetyDepositBox();
